cleaned up the reader by moving dataset analysis to a separate file
